refactor databases out of pymedphys_fileformats
