+++ package.json
@@ -17,7 +17,8 @@
     "netlify:build": "sphinx-build -W docs docs/_build/html",
     "netlify:install": "yarn install:prod && pip install .[docs]",
     "pypi:build": "python setup.py sdist bdist_wheel",
-    "pypi:upload": "twine upload dist/*"
+    "pypi:upload": "twine upload dist/*",
+    "pytest:pylint": "pytest --pylint --pylint-error-types=EF --pylint-jobs=8"
   },
   "husky": {
     "hooks": {
