+++ packages/pymedphys_logfiles/src/pymedphys_logfiles/_install_requires.py
@@ -2,7 +2,6 @@ install_requires = [
     'pymedphys_utilities',
     'pymedphys_fileformats',
     'pymedphys_coordsandscales',
-    'pymedphys_databases',
     'pymedphys_analysis',
     'attrs',
     'numpy < 1.16, >= 1.12',
