+++ packages/pymedphys_logfiles/src/pymedphys_logfiles/elekta/__init__.py
@@ -36,12 +36,16 @@ from pymedphys_utilities.libutils import clean_and_verify_levelled_modules
 
 from ._level1.logfileanalyse import *
 from ._level1.diagnostics_zips import *
-from ._level1.logfileindex import *
+from ._level1.trfidentify import *
+from ._level2.logfileindex import *
 from ._level2.logfilebygantry import *
-from ._level2.orchestration import *
+from ._level3.orchestration import *
 
 clean_and_verify_levelled_modules(globals(), [
-    '._level1.logfileanalyse', '._level1.diagnostics_zips',
-    '._level1.logfileindex', '._level2.logfilebygantry',
-    '._level2.orchestration'
-], package='pymedphys_logfiles.logfile')
+    '._level1.logfileanalyse',
+    '._level1.diagnostics_zips',
+    '._level1.trfidentify',
+    '._level2.logfileindex',
+    '._level2.logfilebygantry',
+    '._level3.orchestration'
+], package='pymedphys_logfiles.elekta')
