+++ packages/pymedphys_logfiles/src/pymedphys_logfiles/elekta/_level2/logfilebygantry.py
