+++ packages/pymedphys_logfiles/tests/logfiles/test_filelogic.py
@@ -25,7 +25,7 @@
 
 
 import os
-from pymedphys_logfiles.logfile import already_indexed_path
+from pymedphys_logfiles.elekta import already_indexed_path
 
 
 def test_file_logic():
