+++ src/pymedphys/trf.py
@@ -1,2 +1,4 @@
 from pymedphys_fileformats.trf import (
-    delivery_data_from_logfile, identify_logfile, decode_header_from_file)
+    delivery_data_from_logfile, decode_header_from_file)
+
+from pymedphys_logfiles.elekta import identify_logfile
